requirements.txt
pytz==2018.3
fitz==0.0.1.dev2
numpy==1.16.2
beautifultable==0.6.0
Pillow==7.1.2
PySimpleGUI27==2.4.1
cx_Freeze==6.1
